fix: rename query types into models query
