+++ packages/bindings-test/src/subspaces/query.rs
@@ -7,7 +7,7 @@ mod tests {
         GroupPermissions, Permission, PermissionDetail, UserPermissions,
     };
     use desmos_bindings::subspaces::query::SubspacesQuery;
-    use desmos_bindings::subspaces::query_types::{
+    use desmos_bindings::subspaces::models_query::{
         QuerySubspaceResponse, QuerySubspacesResponse, QueryUserGroupMembersResponse,
         QueryUserGroupResponse, QueryUserGroupsResponse, QueryUserPermissionsResponse,
     };
