+++ packages/bindings/src/subspaces/models_query.rs
