+++ packages/bindings/src/subspaces/querier.rs
@@ -1,11 +1,11 @@
 //! Contains a querier to query data from the Desmos x/subspaces module.
 
-use crate::subspaces::query_types::{QuerySectionResponse, QuerySectionsResponse};
+use crate::subspaces::models_query::{QuerySectionResponse, QuerySectionsResponse};
 use crate::{
     query::DesmosQuery,
     subspaces::{
         query::SubspacesQuery,
-        query_types::{
+        models_query::{
             QuerySubspaceResponse, QuerySubspacesResponse, QueryUserGroupMembersResponse,
             QueryUserGroupResponse, QueryUserGroupsResponse, QueryUserPermissionsResponse,
         },
