+++ packages/bindings/src/subspaces/query.rs
@@ -11,19 +11,19 @@ use serde::{Deserialize, Serialize};
 #[serde(rename_all = "snake_case")]
 pub enum SubspacesQuery {
     /// Message to query the created subspaces subspaces.
-    /// Response: [QuerySubspacesResponse](crate::subspaces::query_types::QuerySubspacesResponse).
+    /// Response: [QuerySubspacesResponse](crate::subspaces::models_query::QuerySubspacesResponse).
     Subspaces {
         /// Pagination configs.
         pagination: Option<PageRequest>,
     },
     /// Message to query the details of a subspaces.
-    /// Response: [QuerySubspaceResponse](crate::subspaces::query_types::QuerySubspaceResponse).
+    /// Response: [QuerySubspaceResponse](crate::subspaces::models_query::QuerySubspaceResponse).
     Subspace {
         /// Subspace id.
         subspace_id: Uint64,
     },
     /// Message to query the created section inside a subspace.
-    /// Response: [QuerySectionsResponse](crate::subspaces::query_types::QuerySectionsResponse).
+    /// Response: [QuerySectionsResponse](crate::subspaces::models_query::QuerySectionsResponse).
     Sections {
         /// Id of the subspace to query the sections for.
         subspace_id: Uint64,
@@ -31,7 +31,7 @@ pub enum SubspacesQuery {
         pagination: Option<PageRequest>,
     },
     /// Message to query the details of a section.
-    /// Response: [QuerySectionResponse](crate::subspaces::query_types::QuerySectionResponse).
+    /// Response: [QuerySectionResponse](crate::subspaces::models_query::QuerySectionResponse).
     Section {
         /// Id of the subspace to query the sections for.
         subspace_id: Uint64,
@@ -39,7 +39,7 @@ pub enum SubspacesQuery {
         section_id: u32,
     },
     /// Message to query the groups created inside a subspace.
-    /// Response: [QueryUserGroupsResponse](crate::subspaces::query_types::QueryUserGroupsResponse).
+    /// Response: [QueryUserGroupsResponse](crate::subspaces::models_query::QueryUserGroupsResponse).
     UserGroups {
         /// Subspace id to which the groups belongs.
         subspace_id: Uint64,
@@ -49,7 +49,7 @@ pub enum SubspacesQuery {
         pagination: Option<PageRequest>,
     },
     /// Message to query the details of a group.
-    /// Response: [QueryUserGroupResponse](crate::subspaces::query_types::QueryUserGroupResponse).
+    /// Response: [QueryUserGroupResponse](crate::subspaces::models_query::QueryUserGroupResponse).
     UserGroup {
         /// Subspace id to which the group belongs.
         subspace_id: Uint64,
@@ -57,7 +57,7 @@ pub enum SubspacesQuery {
         group_id: u32,
     },
     /// Message to query the member of a group.
-    /// Response: [QueryUserGroupMembersResponse](crate::subspaces::query_types::QueryUserGroupMembersResponse).
+    /// Response: [QueryUserGroupMembersResponse](crate::subspaces::models_query::QueryUserGroupMembersResponse).
     UserGroupMembers {
         /// Subspace id to which the group belongs.
         subspace_id: Uint64,
@@ -68,7 +68,7 @@ pub enum SubspacesQuery {
     },
     /// Message to query the permission that an user have
     /// inside a subspace.
-    /// Response: [QueryUserPermissionsResponse](crate::subspaces::query_types::QueryUserPermissionsResponse).
+    /// Response: [QueryUserPermissionsResponse](crate::subspaces::models_query::QueryUserPermissionsResponse).
     UserPermissions {
         /// Subspace id to which the permissions belongs.
         subspace_id: Uint64,
